add api and validation, message

+++ server/app/controllers/currency.controller.js
@@ -1,3 +1,12 @@
+import {
+    MSG_DELETE_SUCCESSFUL,
+    MSG_ERROR_DELETE,
+    MSG_ERROR_EXISTED,
+    MSG_ERROR_NOT_FOUND,
+    MSG_ERROR_ID_EMPTY,
+    MSG_UPDATE_SUCCESSFUL,
+    MSG_CREATED_SUCCESSFUL
+} from "../utils/message.util";
 import currencyService from "./../services/currency.service";
 import createError from 'http-errors';
 
@@ -5,7 +14,7 @@ exports.findById = async (req, res, next) => {
     try {
         const data = await currencyService.findById(req.params.id);
         if (!data) {
-            return next(createError.BadRequest("Currency not found"));
+            return next(createError.BadRequest(MSG_ERROR_NOT_FOUND("Currency")));
         }
         return res.send({ data });
     } catch (error) {
@@ -35,13 +44,12 @@ exports.createCurrency = async (req, res, next) => {
     try {
         const currencyExisted = await currencyService.findByCurrencyCode(req.body.code);
         if (currencyExisted) {
-            return next(createError.BadRequest("Currency code already exists"));
+            return next(createError.BadRequest(MSG_ERROR_EXISTED("Currency code")));
         }
 
         const data = await currencyService.createCurrency(req.body);
-        return res.send({ message: "Successfully added new currency", data });
+        return res.send({ message: MSG_CREATED_SUCCESSFUL("Currency"), data });
     } catch (error) {
-        console.log(error);
         return next(error);
     }
 }
@@ -50,15 +58,15 @@ exports.updateCurrency = async (req, res, next) => {
     try {
         const currencyExisted = await currencyService.findByCurrencyCode(req.body.code);
         if (currencyExisted && currencyExisted.id !== req.body.currencyId) {
-            return next(createError.BadRequest("Currency code already exists"));
+            return next(createError.BadRequest(MSG_ERROR_EXISTED("Currency code")));
         }
         const foundCurrency = await currencyService.findById(req.body.currencyId);
         if (!foundCurrency) {
-            return next(createError.BadRequest("Currency not found"));
+            return next(createError.BadRequest(MSG_ERROR_NOT_FOUND("Currency")));
         }
 
         await currencyService.updateCurrency(req.body.currencyId, req.body);
-        return res.send({ message: "Successful update" });
+        return res.send({ message: MSG_UPDATE_SUCCESSFUL });
     } catch (error) {
         return next(error);
     }
@@ -67,13 +75,18 @@ exports.updateCurrency = async (req, res, next) => {
 exports.deleteCurrency = async (req, res, next) => {
     try {
         if (!req.params.id && Number(req.params.id)) {
-            return next(createError.BadRequest("CurrencyId cannot be empty"));
+            return next(createError.BadRequest(MSG_ERROR_ID_EMPTY("CurrencyId")));
         }
+        const foundCurrency = await currencyService.findById(req.params.id);
+        if (!foundCurrency) {
+            return next(createError.BadRequest(MSG_ERROR_NOT_FOUND("Currency")));
+        }
+
         await currencyService.deleteCurrency(req.params.id);
-        return res.send({ message: "Successful deletion" });
+        return res.send({ message: MSG_DELETE_SUCCESSFUL });
     } catch (error) {
         return next(
-            createError.BadRequest("This currency cannot be deleted")
+            createError.BadRequest(MSG_ERROR_DELETE("Currency"))
         );
     }
 }
